fix:fix home page & property list pagination
