+++ src/components/PaginationSection.tsx
@@ -1,58 +1,3 @@
-// "use client";
-
-// import {
-//   Pagination,
-//   PaginationContent,
-//   PaginationItem,
-//   PaginationLink,
-//   PaginationNext,
-//   PaginationPrevious,
-// } from "@/components/ui/pagination";
-// import { PaginationMeta } from "@/types/pagination";
-// import { FC } from "react";
-
-// interface PaginationSectionProps extends PaginationMeta {
-//   onChangePage: (page: number) => void;
-// }
-
-// const PaginationSection: FC<PaginationSectionProps> = ({
-//   page,
-//   take,
-//   total,
-//   onChangePage,
-// }) => {
-//   const handlePrev = () => {
-//     if (page > 1) {
-//       onChangePage(page - 1);
-//     }
-//   };
-
-//   const handleNext = () => {
-//     if (page < Math.ceil(total / take)) {
-//       onChangePage(page + 1);
-//     }
-//   };
-//   return (
-//     <Pagination className="mb-12 mt-12 hover:cursor-pointer">
-//       <PaginationContent>
-//         <PaginationItem>
-//           <PaginationPrevious onClick={handlePrev} />
-//         </PaginationItem>
-
-//         <PaginationItem>
-//           <PaginationLink>{page}</PaginationLink>
-//         </PaginationItem>
-
-//         <PaginationItem>
-//           <PaginationNext onClick={handleNext} />
-//         </PaginationItem>
-//       </PaginationContent>
-//     </Pagination>
-//   );
-// };
-
-// export default PaginationSection;
-
 "use client";
 
 import {
@@ -63,23 +8,19 @@ import {
   PaginationNext,
   PaginationPrevious,
 } from "@/components/ui/pagination";
+import { PaginationMeta } from "@/types/pagination";
 import { FC } from "react";
 
-interface PaginationSectionProps {
-  page: number;
-  take: number;
-  totalCount: number;
+interface PaginationSectionProps extends PaginationMeta {
   onChangePage: (page: number) => void;
 }
 
 const PaginationSection: FC<PaginationSectionProps> = ({
   page,
   take,
-  totalCount,
+  total,
   onChangePage,
 }) => {
-  const totalPages = Math.ceil(totalCount / take);
-
   const handlePrev = () => {
     if (page > 1) {
       onChangePage(page - 1);
@@ -87,112 +28,23 @@ const PaginationSection: FC<PaginationSectionProps> = ({
   };
 
   const handleNext = () => {
-    if (page < totalPages) {
+    if (page < Math.ceil(total / take)) {
       onChangePage(page + 1);
     }
   };
-
-  // Don't render pagination if there's only one page
-  if (totalPages <= 1) {
-    return null;
-  }
-
   return (
-    <Pagination className="mb-12 mt-12">
+    <Pagination className="mb-12 mt-12 hover:cursor-pointer">
       <PaginationContent>
         <PaginationItem>
-          <PaginationPrevious
-            onClick={handlePrev}
-            className={
-              page <= 1 ? "pointer-events-none opacity-50" : "cursor-pointer"
-            }
-          />
+          <PaginationPrevious onClick={handlePrev} />
         </PaginationItem>
 
-        {totalPages <= 7 ? (
-          // Show all pages if 7 or fewer
-          [...Array(totalPages)].map((_, i) => (
-            <PaginationItem key={i + 1}>
-              <PaginationLink
-                onClick={() => onChangePage(i + 1)}
-                isActive={page === i + 1}
-                className="cursor-pointer"
-              >
-                {i + 1}
-              </PaginationLink>
-            </PaginationItem>
-          ))
-        ) : (
-          // Show first, last, current and some surrounding pages
-          <>
-            {/* First page */}
-            <PaginationItem>
-              <PaginationLink
-                onClick={() => onChangePage(1)}
-                isActive={page === 1}
-                className="cursor-pointer"
-              >
-                1
-              </PaginationLink>
-            </PaginationItem>
-
-            {/* Ellipsis if needed */}
-            {page > 3 && (
-              <PaginationItem>
-                <PaginationLink className="cursor-default">...</PaginationLink>
-              </PaginationItem>
-            )}
-
-            {/* Pages around current */}
-            {[...Array(3)]
-              .map((_, i) => {
-                const pageNum = Math.max(2, page - 1) + i;
-                if (pageNum > 1 && pageNum < totalPages) {
-                  return (
-                    <PaginationItem key={pageNum}>
-                      <PaginationLink
-                        onClick={() => onChangePage(pageNum)}
-                        isActive={page === pageNum}
-                        className="cursor-pointer"
-                      >
-                        {pageNum}
-                      </PaginationLink>
-                    </PaginationItem>
-                  );
-                }
-                return null;
-              })
-              .filter(Boolean)}
-
-            {/* Ellipsis if needed */}
-            {page < totalPages - 2 && (
-              <PaginationItem>
-                <PaginationLink className="cursor-default">...</PaginationLink>
-              </PaginationItem>
-            )}
-
-            {/* Last page */}
-            <PaginationItem>
-              <PaginationLink
-                onClick={() => onChangePage(totalPages)}
-                isActive={page === totalPages}
-                className="cursor-pointer"
-              >
-                {totalPages}
-              </PaginationLink>
-            </PaginationItem>
-          </>
-        )}
+        <PaginationItem>
+          <PaginationLink>{page}</PaginationLink>
+        </PaginationItem>
 
         <PaginationItem>
-          <PaginationNext
-            onClick={handleNext}
-            className={
-              page >= totalPages
-                ? "pointer-events-none opacity-50"
-                : "cursor-pointer"
-            }
-          />
+          <PaginationNext onClick={handleNext} />
         </PaginationItem>
       </PaginationContent>
     </Pagination>
